删除角色

+++ src/main/java/com/anthonyzero/scaffold/system/controller/UserController.java
@@ -166,6 +166,9 @@ public class UserController extends BaseController {
     @GetMapping("delete/{userIds}")
     @RequiresPermissions("user:delete")
     public Response deleteUsers(@PathVariable String userIds) {
+        if (StringUtils.isBlank(userIds)) {
+            return Response.error(CodeMsgEnum.PARAMETER_ERROR);
+        }
         String[] ids = userIds.split(StringPool.COMMA);
         this.userService.deleteUsers(ids);
         return Response.success();
